requirements.txt
# Core Dependencies
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.0.20
langchain-community==0.0.13

# Data Collection
fredapi==0.5.1
yfinance==0.2.36
pandas==2.1.4
numpy==1.26.2

# Analysis & Forecasting
scikit-learn==1.3.2
statsmodels==0.14.1
prophet==1.1.5
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
certifi  # SSL certificate handling for macOS

# Optional but recommended
jupyter==1.0.0
ipython==8.19.0